To Buy and Calculator Bug fix
